+++ CHANGELOG.md
@@ -5,6 +5,10 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [Unreleased]
+### Added
+- Add `browser-sync` with command `npm run dev` to support live-reload
+
 ## [1.2.2] 2025-08-30
 ### Fixed
 - Fix missing cursor pointer on site log and sun/moon